Find the index of first 1 in a sorted array of 0’s and 1’s
Given a sorted array consisting 0’s and 1’s. The problem is to find the index of first ‘1’ in the sorted array. It could be possible that the array consists of only 0’s or only 1’s. If 1’s are not present in the array then print “-1”.

Examples :

Input : arr[] = {0, 0, 0, 0, 0, 0, 1, 1, 1, 1}
Output : 6
The index of first 1 in the array is 6.

Input : arr[] = {0, 0, 0, 0}
Output : -1
1's are not present in the array.
Source: Asked in Amazon Interview

Recommended: Please solve it on “PRACTICE ” first, before moving on to the solution.
Naive Approach: Traverse the array from left to right and return the index of first ‘1’. If 1’s are not present in the array, then print “-1”.