Position of rightmost different bit
Given two numbers m and n. Find the position of rightmost different bit in binary representation of numbers. It is guaranteed that such a bit exists.

Examples :

Input : m = 11, n = 9
Output : 2
(11)10 = (1011)2
(9)10 = (1001)2
It can be seen that 2nd bit from
the right is different 

Input : m = 52, n = 4
Output : 5
(52)10 = (110100)2
(4)10 = (100)2, can also be written as
     = (000100)2
It can be seen that 5th bit from
the right is different            
Recommended: Please solve it on “PRACTICE” first, before moving on to the solution.
Approach: Get the bitwise xor of m and n. Let it be xor_value = m ^ n. Now, find the position of rightmost set bit in xor_value.



 

Explanation: The bitwise xor operation produces a number which has set bits only at the positions where the bits of m and n differ. Thus, the position of rightmost set bit in xor_value gives the position of rightmost different bit.